fix(frontend): updated tx modal logic
